GetTickCount() -> GetTickCount64()
    //上改成下
    static DWORD a;
    static ULONGLONG a;
    //上改成下
    a = GetTickCount();
    a = GetTickCount64();
    //上改成下
    wsprintf(a, L"%u", GetTickCount64());
    wsprintf(a, L"%llu", GetTickCount64());
    //上改成下
    DWORD a(){}
    ULONGLONG a(){}

+++ dll/taskbar_inspector.c
@@ -155,7 +155,7 @@ void CloseInspectorDlg()
 	}
 }
 
-DWORD GetInspectorCloseTime()
+ULONGLONG GetInspectorCloseTime()
 {
 	return dwInspectorCloseTime;
 }
